ollama-client/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
aiohttp>=3.9.1
websockets>=12.0
python-dotenv>=1.0.1
python-dateutil>=2.8.2
psutil>=5.9.0
py-cpuinfo>=9.0.0
nvidia-ml-py>=12.535.108
torch>=2.1.0
